+++ packages/nestjs/src/index.ts
@@ -2,6 +2,6 @@ export * from '@sentry/node';
 
 export { init } from './sdk';
 
-export { SentryTraced } from './span-decorator';
-export { SentryCron } from './cron-decorator';
-export { WithSentry } from './error-decorator';
+export { SentryTraced } from './decorators/sentry-traced';
+export { SentryCron } from './decorators/sentry-cron';
+export { WithSentry } from './decorators/with-sentry';
